"Updated MethodTransformer to handle 'typeSwitch' in Java 21"
In the MethodTransformer.java file, an update was made to handle a new construct introduced in Java 21. Now, when invoking dynamic methods and checking for a 'typeSwitch' method specifically, the logger records an integer return type instead of an object as previously done. This leverages the feature of 'typeSwitch' method in new Java 21's switch case.